Finished off changes to convert ino to cpp
